+++ src/main/java/com/easypost/model/PayloadCollection.java
@@ -1,0 +1,10 @@
+package com.easypost.model;
+
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class PayloadCollection {
+    private List<Payload> payloads;
+}