+++ src/main/java/com/easypost/service/EventService.java
@@ -1,5 +1,6 @@
 package com.easypost.service;
 
+import java.util.List;
 import java.util.Map;
 
 import com.easypost.exception.EasyPostException;
@@ -7,6 +8,8 @@ import com.easypost.http.Requestor;
 import com.easypost.http.Requestor.RequestMethod;
 import com.easypost.model.Event;
 import com.easypost.model.EventCollection;
+import com.easypost.model.Payload;
+import com.easypost.model.PayloadCollection;
 import com.easypost.utils.InternalUtilities;
 
 public class EventService {
@@ -43,5 +46,35 @@ public class EventService {
     public EventCollection all(final Map<String, Object> params) throws EasyPostException {
         return Requestor.request(RequestMethod.GET, InternalUtilities.classURL(Event.class), params,
                 EventCollection.class, client);
+    }
+
+    /**
+     * Retrieve all payloads.
+     *
+     * @param eventId The ID of event.
+     * @return List of Payload objects.
+     * @throws EasyPostException when the request fails.
+     */
+    public List<Payload> retrieveAllPayloads(final String eventId) throws EasyPostException {
+        PayloadCollection payloads = Requestor.request(RequestMethod.GET,
+                String.format("%s/%s", InternalUtilities.instanceURL(Event.class, eventId), "payloads"), null,
+                PayloadCollection.class, client);
+
+        return payloads.getPayloads();
+    }
+
+    /**
+     * Retrieve a payload.
+     *
+     * @param eventId   The ID of event.
+     * @param payloadId The ID of payload.
+     * @return PayloadCollection object
+     * @throws EasyPostException when the request fails.
+     */
+    public Payload retrievePayload(final String eventId, final String payloadId) throws EasyPostException {
+        return Requestor.request(RequestMethod.GET,
+                String.format("%s/%s/%s", InternalUtilities.instanceURL(Event.class, eventId), "payloads", payloadId),
+                null,
+                Payload.class, client);
     }
 }